requirements.txt
gevent-websocket==0.10.1
PyYAML==5.4.1